listening to peer ports for bin coms implemented

+++ src/modules/constants.rs
@@ -1,37 +1,37 @@
 pub mod constants {
     pub struct Constants{
         pub nodeversion: String,
-        pub default_chunk: i32,
+        pub default_chunk: u32,
         // These bytes are carefully selected to make the base58 and base32 representations of different CID types
         // easy to distinguish and not collide with anything on https://github.com/multiformats/multicodec
-        pub cid_type_raw: i32,
-        pub cid_type_metadata_file: i32,
-        pub cid_type_metadata_dir: i32,
-        pub cid_type_resolver: i32,
+        pub cid_type_raw: u8,
+        pub cid_type_metadata_file: u8,
+        pub cid_type_metadata_dir: u8,
+        pub cid_type_resolver: u8,
         
-        pub registry_s5_magic_byte: i32,
-        pub metadata_magic_byte: i32,
+        pub registry_s5_magic_byte: u8,
+        pub metadata_magic_byte: u8,
         
         // types for metadata files
-        pub metadata_type_file: i32,
-        pub metadata_type_chunked_file: i32,
-        pub metadata_type_directory: i32,
+        pub metadata_type_file: u8,
+        pub metadata_type_chunked_file: u8,
+        pub metadata_type_directory: u8,
         
-        pub registry_max_data_size: i32,
+        pub registry_max_data_size: u8,
         
-        pub mhash_blake3: [i32;2],
+        pub mhash_blake3: [u8;2],
         
         pub mkey_ed25519: u8,
         
         //  Use this for protocol updates
-        pub protocol_method_handshake_open: i32,
-        pub protocol_method_handshake_done: i32,
-        pub protocol_methods_signed_message: i32,
-        pub protocol_method_hash_query_response: i32,
-        pub protocol_method_hash_query: i32,
-        pub protocol_method_annouce_peers: i32,
-        pub protocol_method_registry_update: i32,
-        pub protocol_method_registry_query: i32,
+        pub protocol_method_handshake_open: u8,
+        pub protocol_method_handshake_done: u8,
+        pub protocol_methods_signed_message: u8,
+        pub protocol_method_hash_query_response: u8,
+        pub protocol_method_hash_query: u8,
+        pub protocol_method_annouce_peers: u8,
+        pub protocol_method_registry_update: u8,
+        pub protocol_method_registry_query: u8,
     }
     impl Constants{
         pub fn get_constants() -> Constants {
